# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/{type}/{id}/image:
    get:
      tags:
      - image-api-controller
      summary: Gets an entity image.
      operationId: getImage
      parameters:
      - name: type
        in: path
        description: Type of entity (either user or post).
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: Entity associated identifier (License ID for users and ID for
          posts).
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Image was not properly processed.
        "404":
          description: Image could not be found (because entity instance does not
            exist or it does not have associated image).
        "403":
          description: Current logged user is not authorized to download this image.
        "200":
          description: Image could be found and is downloaded.
          content:
            image/*: {}
    put:
      tags:
      - image-api-controller
      summary: Changes the image of a user or a post by a new one and updates the
        entity instance.
      operationId: uploadImage
      parameters:
      - name: type
        in: path
        description: Type of entity (either user or post).
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: Entity associated identifier (License ID for users and ID for
          posts).
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  description: Multipart File.
                  format: binary
      responses:
        "200":
          description: Image could be found and is downloaded.
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/User'
                - $ref: '#/components/schemas/Post'
        "400":
          description: Image was not properly processed.
        "403":
          description: Current logged user is not authorized to upload or change this
            image.
...
  /api/posts/{id}:
    get:
      tags:
      - post-api-controller
      summary: Get a simple post.
      operationId: newsPost
      parameters:
      - name: id
        in: path
        description: Unique identifier of a post.
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Resource not found.
        "200":
          description: Found the post.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    put:
      tags:
      - admin-post-api-controller
      summary: Updates a post.
      operationId: updatePost
      parameters:
      - name: id
        in: path
        description: ID of post.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDTO'
        required: true
      responses:
        "500":
          description: Post cannot be modified on the basis of failed data.
        "200":
          description: Edit the post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
    delete:
      tags:
      - admin-post-api-controller
      summary: Deletes a post.
      operationId: deletePost
      parameters:
      - name: id
        in: path
        description: Post id.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Elimination successfully completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "404":
          description: Request is invalid because of empty or non-existent post retrieve.
...
  /api/competitions/{id}/control:
    put:
      tags:
      - referee-control-api-controller
      summary: Saves the result of a competition.
      operationId: controlCompetition
      parameters:
      - name: id
        in: path
        description: ID of the competition.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionResultDTO'
        required: true
      responses:
        "200":
          description: Save successfully completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionResultDTO'
        "404":
          description: Request is invalid because of non-finished competition retrieve.
...
  /api/posts:
    get:
      tags:
      - post-api-controller
      summary: Get a list with the posts to view (paginated).
      operationId: getPage
      parameters:
      - name: page
        in: query
        description: Page number.
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        description: Size of the post list to show.
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Page with more than one post to view.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        "400":
          description: Request is invalid because of empty or non-existent page retrieve.
    post:
      tags:
      - admin-post-api-controller
      summary: Creates a new post.
      operationId: addNewPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDTO'
        required: true
      responses:
        "200":
          description: Creation of a new post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        "500":
          description: Post cannot be created on the basis of failed data.
...
  /competitions/{idCompetition}/control:
    get:
      tags:
      - referee-control-controller
      operationId: controlCompetition_1
      parameters:
      - name: idCompetition
        in: path
        required: true
        schema:
          type: string
      - name: winner
        in: query
        required: true
        schema:
          type: string
      - name: loser
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
...
components:
  schemas:
    User:
      type: object
      properties:
        licenseId:
          type: string
        name:
          type: string
        surname:
          type: string
        gender:
          type: string
          enum:
          - H
          - M
        phone:
          type: integer
          format: int32
        email:
          type: string
        dni:
          type: string
        nickname:
          type: string
        securityQuestion:
          type: string
        securityAnswer:
          type: string
        belt:
          type: string
          enum:
          - B
          - BAm
          - Am
          - AmN
          - "N"
          - NV
          - V
          - VAz
          - Az
          - AzM
          - M
          - N1
          - N2
          - N3
          - N4
          - N5
          - N6
          - N7
          - N8
          - N9
          - N10
        gym:
          type: string
        weight:
          type: integer
          format: int32
        refereeRange:
          type: string
          enum:
          - S
          - E
          - C
          - A
        roles:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
            - C
            - R
            - A
        imageFile:
          type: string
        birthdate:
          type: string
    Post:
      type: object
      properties:
        idPost:
          type: integer
          format: int32
        author:
          $ref: '#/components/schemas/User'
        title:
          type: string
        body:
          type: string
        timestamp:
          type: string
          format: date-time
        formattedEditionTimestamp:
          type: string
        imageFile:
          type: string
    PostDTO:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
      description: Post DTO.
...
    CompetitionResultDTO:
      type: object
      properties:
        idCompetition:
          type: string
        winner:
          type: string
        loser:
          type: string
...